updates labels on existing SourceForge API v2.0 Beta projects
This script will update the labels attached to existing issues in
the SourceForge v2.0 API Beta.  This is mildly tricky because the API
clobbers existing labels.  Our basic operation is thus to fetch the
existing labels, extend the list with new ones of interest, and then
assign the new labels.